Wizard: edit Security review step according to mocks (HMS-9597)
This edits the review step for Security according to the UI mocks.
